traits for dedekind domain
