+++ rings/src/structure/mod.rs
@@ -2,8 +2,10 @@ mod elements;
 mod factorization;
 mod quotient;
 mod structure;
+mod ideals;
 
 pub use elements::*;
 pub use factorization::*;
 pub use quotient::*;
 pub use structure::*;
+pub use ideals::*;
